requirements.txt
requests
beautifulsoup4
supabase
lxml
python-dotenv
mutagen
feedgen
pytest
pytest-cov
ruff
